requirement.txt
baresipy
RPi.GPIO
python-dotenv
sounddevice